Copy project.yml and .env files from src
This allows users to write their own project.yml with more features such as environment variables and triggers, and that just gets copied to the build rather than a newly generated minimal project.yml.
If the user doesn't provide one, it still gets generated.
If the user provides a partial project.yml, the missing functions will be added.
Also checks for the existence of a .env file and copies that to the build dir.
Both new file locations can be overridden with command line flags.

Signed-off-by: Brian Evans <ebrian101@gmail.com>